refactor(agent-tars-server): cli
